print binary rep of number

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -12,7 +12,7 @@ unsigned int power(unsigned int x, unsigned int y);
 
 void print_binary(unsigned long int n)
 {
-	unsigned int i = 0, max_power, num = n;
+	unsigned long int i = 0, max_power, num = n;
 
 	if (n == 0)
 	{
